// FRONTEND/chatbot-admin-panel/src/services/api/schemas/roleResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { RoleResponseDescription } from './roleResponseDescription';

export interface RoleResponse {
  /**
   * Nombre único del rol (ej. superadmin, context_editor)
   * @minLength 3
   * @maxLength 50
   */
  name: string;
  /** Descripción detallada del rol. */
  description?: RoleResponseDescription;
  id: number;
  created_at: string;
  updated_at: string;
}