// FRONTEND/chatbot-admin-panel/src/services/api/schemas/appUserLocalCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { AppUserLocalCreateEmail } from './appUserLocalCreateEmail';
import type { AppUserLocalCreateFullName } from './appUserLocalCreateFullName';

/**
 * Schema para crear un nuevo usuario LOCAL a través de la API.
 */
export interface AppUserLocalCreate {
  /**
   * @minLength 3
   * @maxLength 100
   */
  username_ad: string;
  email?: AppUserLocalCreateEmail;
  full_name?: AppUserLocalCreateFullName;
  is_active_local?: boolean;
  /**
   * Contraseña requerida para el nuevo usuario.
   * @minLength 8
   */
  password: string;
  /** Lista de IDs de los roles a asignar. */
  role_ids?: number[];
}